Fix final frame in animation
If an animation stops at the last frame, it might be wrapped to the first frame.
This can be fixed by adding a small value to the duration (1e-6).

The common code in GVRSkinningController and GVRNodeAnimationController is pulled
into a base class, and the fixed is applied on that.

Rename some variables to distinguish time vs ticks in the animation controller(s).
