+++ GVRf/Framework/src/org/gearvrf/animation/keyframe/GVRNodeAnimationController.java
@@ -9,11 +9,9 @@ import org.joml.Matrix4f;
 /**
  * Controls node animation.
  */
-public class GVRNodeAnimationController {
+public class GVRNodeAnimationController extends GVRAnimationController {
     private static final String TAG = GVRNodeAnimationController.class.getSimpleName();
-
     protected GVRSceneObject sceneRoot;
-    protected GVRKeyFrameAnimation animation;
 
     protected class AnimationItem {
         GVRSceneObject target;
@@ -35,8 +33,8 @@ public class GVRNodeAnimationController {
      * @param animation The animation object.
      */
     public GVRNodeAnimationController(GVRSceneObject sceneRoot, GVRKeyFrameAnimation animation) {
+        super(animation);
         this.sceneRoot = sceneRoot;
-        this.animation = animation;
 
         animatedNodes = new ArrayList<AnimationItem>();
         if (animation != null) {
@@ -62,22 +60,13 @@ public class GVRNodeAnimationController {
     }
 
     /**
-     * Update node transforms at each animation step.
+     * Update node transforms to a tick.
+     * @param animationTick
+     *         The tick to animate to.
      */
-    public void animate(float timeInSeconds) {
-        float ticksPerSecond;
-        float timeInTicks;
-        Matrix4f[] animationTransform = null;
-
-        if (animation.mTicksPerSecond != 0) {
-            ticksPerSecond = (float) animation.mTicksPerSecond;
-        } else {
-            ticksPerSecond = 25.0f;
-        }
-        timeInTicks = timeInSeconds * ticksPerSecond;
-
-        float animationTime = timeInTicks % animation.mDurationTicks; // auto-repeat
-        animationTransform = animation.getTransforms(animationTime);
+    @Override
+    protected void animateImpl(float animationTick) {
+        Matrix4f[] animationTransform = animation.getTransforms(animationTick);
 
         for (AnimationItem item : animatedNodes) {
             item.target.getTransform().setModelMatrix(animationTransform[item.channelId]);
